Migrate empty table types and utils to shared location
